fix test_all_types.py
