+++ tools/pythonpkg/tests/fast/test_all_types.py
@@ -7,7 +7,39 @@ from decimal import Decimal
 from uuid import UUID
 import pytz
 import pytest
-from conftest import recursive_equality
+
+
+def replace_with_ndarray(obj):
+    if hasattr(obj, '__getitem__'):
+        if isinstance(obj, dict):
+            for key, value in obj.items():
+                obj[key] = replace_with_ndarray(value)
+        elif isinstance(obj, list):
+            for i, item in enumerate(obj):
+                obj[i] = replace_with_ndarray(item)
+        return np.array(obj)
+    return obj
+
+
+# we need to write our own equality function that considers nan==nan for testing purposes
+def recursive_equality(o1, o2):
+    import math
+
+    if type(o1) != type(o2):
+        return False
+    if type(o1) == float and math.isnan(o1) and math.isnan(o2):
+        return True
+    if o1 is np.ma.masked and o2 is np.ma.masked:
+        return True
+    try:
+        if len(o1) != len(o2):
+            return False
+        for i in range(len(o1)):
+            if not recursive_equality(o1[i], o2[i]):
+                return False
+        return True
+    except:
+        return o1 == o2
 
 
 # Regenerate the 'all_types' list using:
